api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scholarly==1.7.11
google-genai==1.38.0
pytz==2023.3
